+++ stdlib/VERSIONS
@@ -34,6 +34,7 @@ _dummy_thread: 3.0-3.8
 _dummy_threading: 2.7-3.8
 _heapq: 2.7-
 _imp: 3.0-
+_io: 3.0-
 _json: 2.7-
 _locale: 2.7-
 _markupbase: 2.7-
