+++ tests/stubtest_allowlists/py3_common.txt
@@ -52,17 +52,6 @@ importlib.abc.Loader.exec_module  # See Lib/importlib/_abc.py. Might be defined 
 importlib.abc.MetaPathFinder.find_spec  # Not defined on the actual class, but expected to exist.
 importlib.abc.PathEntryFinder.find_spec  # Not defined on the actual class, but expected to exist.
 importlib.machinery.ExtensionFileLoader.get_filename  # Wrapped with _check_name decorator which changes runtime signature
-io.BufferedRandom.truncate
-io.BufferedReader.seek
-io.BufferedReader.truncate
-io.BufferedWriter.seek
-io.BufferedWriter.truncate
-io.BytesIO.readlines
-io.BytesIO.seek  # Parameter name for a positional-only param differs from its name in the inherited method
-io.FileIO.seek
-io.StringIO.seek
-io.StringIO.truncate
-io.TextIOWrapper.truncate
 ipaddress.IPv4Interface.hostmask
 ipaddress.IPv6Interface.hostmask
 ipaddress._BaseNetwork.broadcast_address
